pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=70.0.0"]

[project]
authors = [
    { name = "Gabriel Gutierrez", email = "gabriel.bgs00@gmail.com" },
    { name = "Otávio Napoli", email = "otavio.napoli@ic.unicamp.br" },
    { name = "Fernando Gubitoso Marques", email = "fernandogubitosom@gmail.com" },
    { name = "Edson Borin", email = "borin@unicamp.br" },

]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
description = "A framework for training machine learning models for researchers."
keywords = [
    "Deep Learning",
    "Machine Learning",
    "Pytorch",
    "Research",
    "Self-Supervised Learning",
    "SSL",
]
license = { file = "LICENSE" }
name = "minerva-ml"
readme = "README.md"
requires-python = ">=3.10"
version = "0.3.8-beta"

dependencies = [
    "gitpython>=3",
    "jsonargparse[omegaconf]==4.37.0",
    "kaleido>=0.2.1",
    "lightning>=2.1.9",
    "matplotlib>=3.7",
    "numpy>=1.23.5",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.2",
    "perlin-noise>=1.12",
    "plotly>=5.18",
    "PyYAML>=6.0",
    "rich>=13.0",
    "scikit-learn>=1.2.0",
    "scipy>=1.10",
    "statsmodels>=0.13",
    "tifffile>=2024",
    "timm>=1.0.2",
    "torch>=2.5.0",
    "torchinfo>=1.8.0",
    "torchmetrics>=1.3.0",
    "torchvision>=0.15",
    "transformers>=4.47.1",
    "wrapt>=1.16",
    "zarr>=2.17",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["minerva*"]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "interrogate",
    "ipykernel",
    "mock",
    "nbformat",
    "pytest-coverage",
]
docs = [
    "Ipython",
    "myst-parser",
    "nbsphinx",
    "pandoc",
    "sphinx-argparse",
    "sphinx-autoapi",
    "sphinx-autodoc-typehints",
    "sphinx-rtd-theme",
    "sphinx",
]
analysis = ["codecarbon>=2.8.3", "ultralytics-thop>=2.0.14"]

[project.urls]
"Bug Tracker" = "https://github.com/discovery-unicamp/Minerva/issues"
"Homepage" = "https://github.com/discovery-unicamp/Minerva"
